fix(widgets): one place that encodes a colour, and it premultiplies
`linear_to_srgb_u8` had three copies and `paint_tint_region` two, and
they had already drifted in a way that mattered: `shape` premultiplied
before the sRGB encode and the two text widgets did not.

That is a bug, not a stylistic difference. The pipeline blends with
PREMULTIPLIED_ALPHA_BLENDING, so a straight-alpha tint composites too
bright. It stayed invisible because text is almost always painted opaque,
where the two agree — but a translucent `Button` label or a `RichText`
span colour was wrong, and the fix could not reach them from the copy
that had it.

The new `color` module is that reachable place. `shape::tint_region`
keeps its cache and now allocates through the same function, so the
1x1-region upload exists once rather than three times.

Visual check wanted on the showcase's decoration and controls sections:
anything painted with alpha < 1 should now read darker, which is correct.
