+++ matcha-ecs-widgets/src/text.rs
@@ -185,56 +185,9 @@ pub(crate) fn shape(font_ctx: &FontCtx, content: &str, font_size: f32, max_width
     font_ctx.0.font_system.layout_text(&data, &config)
 }
 
-/// Gamma-encode a linear colour component into the sRGB space the atlas
-/// texture is stored in (matches what a render pass writing to an
-/// `Rgba8UnormSrgb` target does automatically; `write_data` is a raw byte
-/// copy with no such conversion, so it must be done by hand here).
-fn linear_to_srgb_u8(c: f32) -> u8 {
-    let c = c.clamp(0.0, 1.0);
-    let encoded = if c <= 0.0031308 {
-        c * 12.92
-    } else {
-        1.055 * c.powf(1.0 / 2.4) - 0.055
-    };
-    (encoded * 255.0).round() as u8
-}
-
-/// Paint a 1x1 solid-colour region into `ctx.texture_atlas`, reused
-/// (UV-clamped to any on-screen size) as every glyph's stencil "tint".
-///
-/// Written directly via `write_data` rather than `ColorRect`'s render-pass +
-/// `VertexColor` approach: a real GPU render pass whose viewport is scoped to
-/// a 1x1 (or otherwise very small, single-digit-pixel) atlas region was found
-/// to rasterise incorrectly (a soft, mispositioned blob instead of a flat
-/// fill — reproduced in isolation with a hand-built 4x4 case, unrelated to
-/// glyphs/stencils). `ColorRect` never hits this because it always sizes
-/// regions to its own (typically much larger) rect, so the bug went
-/// unnoticed; text's single shared tint pixel triggers it directly. Root
-/// cause not identified (deferred — see `ECS_IMPLEMENTATION_PLAN.md` §8);
-/// `write_data` sidesteps it entirely and is a strictly simpler upload for a
-/// flat fill anyway.
+/// Paint the 1x1 tint pixel every glyph's stencil is masked against.
 pub(crate) fn paint_tint_region(ctx: &RenderCtx, color: [f32; 4]) -> Option<AtlasRegion> {
-    let region = match ctx.texture_atlas.allocate(ctx.device, ctx.queue, [1, 1]) {
-        Ok(region) => region,
-        Err(e) => {
-            log::error!("Text tint region allocation failed: {e}");
-            return None;
-        }
-    };
-
-    let alpha = color[3].clamp(0.0, 1.0);
-    let bytes = [
-        linear_to_srgb_u8(color[0]),
-        linear_to_srgb_u8(color[1]),
-        linear_to_srgb_u8(color[2]),
-        (alpha * 255.0).round() as u8,
-    ];
-    if let Err(e) = region.write_data(ctx.queue, &bytes) {
-        log::error!("Text tint upload failed: {e}");
-        return None;
-    }
-
-    Some(region)
+    crate::color::paint_tint_region(ctx, color, "Text")
 }
 
 /// Composite `layout`'s glyphs into `(node, local_translation)` pairs, each a
